+++ lib/super_sliver_list.dart
@@ -1,7 +1,8 @@
 export "src/extent_manager.dart" show ItemRange;
+export "src/header_anchor.dart";
 export "src/layout_budget.dart";
-export "src/perf_flags.dart";
 export "src/list_view.dart";
+export "src/perf_flags.dart";
 export "src/scroll_physics.dart";
 export "src/sliver_extensions.dart";
 export "src/stick_target.dart";